requirements.txt
python-dotenv
psutil
requests